lib/ptexter/src/requirements-wc-l.txt
# Requirements for wc-l endpoint - high performance line counting
polars-lts-cpu==1.30.0
